sensors: Make trigger/events work for android-iio-hal
android-iio-hal can make use of trigger/events to read from sensors
more efficiently (for example on when actually moving the device).
To make that work in needs write access to several sysfs files.
Unfortunately the ueventd.rc does not seem to support wildcards for the
filenames, so this just adds the relevant files for idol347 right now.

This allows dropping the android-iio-hal quirks for everything but the
magnetometer. The magnetometer in idol347 does not have an interrupt
and can therefore not make use of the trigger functionality (not sure
why android-iio-hal cannot detect this automatically).